backend/simple_services/event/requirements.txt
fastapi>==0.110.0
pydantic==2.6.4
uvicorn==0.29.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28
boto3==1.34.74
python-multipart==0.0.9

